requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
feedparser>=6.0.0
